pyproject.toml
[tool.poetry]
name = "python_in_science"
version = "0.1.0"
description = ""
authors = ["Dawid Karpiński"]

[tool.poetry.dependencies]
python = "^3.11"
ascii-graph = "^1.5.1"
tqdm = "^4.66.5"
pillow = "^11.0.0"
numpy = "^2.0.0"
numba = "^0.60.0"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"
selenium = "^4.27.1"
webdriver-manager = "^4.0.2"
matplotlib = "^3.10.0"
scipy = "^1.14.1"
scienceplots = "^2.1.1"
sympy = "^1.13.3"
bokeh = "^3.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
